fix(render): VDM 段偏移 push constant 每 DrawBatch 必推 + 索引数组 U32 化残留
- PushConstants 移出 need_buffer_switch（VDM 共享 buffer 只触发一次——后续模型
  段偏移未推——读错段顶点坍缩；每 DrawBatch 必推）
- PlaneAndSquare/AxisBoundingBoxSquareArray 的 uint16 索引数组改 uint32
  （WriteIBO 按 IBO 类型读数据——U32 IBO 读 uint16 数组错位——索引超范围）

+++ src/InlineGeometry/PlaneAndSquare.cpp
@@ -124,7 +124,7 @@ namespace hgl::graph::inline_geometry
                 float           xy_tex_coord[] = {  0.0f, 0.0f,        1.0f, 0.0f,          1.0f, 1.0f,          0.0f, 1.0f        };
         const   Vector3f  xy_normal(0.0f,0.0f,1.0f);
         const   Vector3f  xy_tangent(1.0f,0.0f,0.0f);
-        const   uint16          indices[]={0,1,2,0,2,3};
+        const   uint32          indices[]={0,1,2,0,2,3};
 
         if(!pc)return(nullptr);
